Cobot now detects distance and changes it in the observable collection

+++ Valkwelding Kunststof Lassen/Valkwelding Kunststof Lassen/Touch PoC/Configuration/CobotSettings.cs
@@ -10,7 +10,7 @@ namespace ValkWelding.Welding.Touch_PoC.Configuration
     {
         public string IpAddress { get; set; }
         public int Port { get; set; }
-        public int MovementSpeed { get; set; }
-        public int MovementStepSize { get; set; }
+        public float MovementSpeed { get; set; }
+        public float MovementStepSize { get; set; }
     }
 }
